snapshot

+++ bash/db-create-vitriol.sh
@@ -1,23 +1,54 @@
+function create_schema {
+
+  usage="create_schema [database]..."
+
+  [ -z "$1" ] && echo $usage && return
+
+    echo 'creating $1 db...'
+    mysql -u root < $MILDRED_HOME/db/design/$1.sql
+}
+
+function regenerate_alchemy_classes {
+
+  usage="regenerate_alchemy_classes [database]..."
+
+  [ -z "$1" ] && echo $usage && return
+    # echo 'creating $1 db...'
+    sqlacodegen mysql://root:steel@localhost/$1 > $MILDRED_HOME/python/server/db/generated/sqla_$1.py
+}
+
 clear
-rm $MILDRED_HOME/python/server/db/generated/*.p*
 
-echo 'updating media'
-mysql cuba  < $MILDRED_HOME/db/design/media.sql
-echo 'updating media action'
-mysql cuba < $MILDRED_HOME/db/design/analysis.sql
-echo 'updating media introspection'
-mysql cuba < $MILDRED_HOME/db/design/service.sql
-echo 'updating scratch'
-mysql cuba < $MILDRED_HOME/db/design/scratch.sql
+cd $MILDRED_HOME/bash
 
+echo 'dropping schemas...'
+mysql -u root < $MILDRED_HOME/db/design/drop-all.sql
+
+echo 'creating schemas...'
+
+# create_schema admin
+# create_schema elastic
+# create_schema analysis
+# create_schema media
+# create_schema service
+# create_schema suggestion
+# create_schema mildred
+# create_schema scratch
+
+mysql -u root < $MILDRED_HOME/db/design/all-databases.sql
+
+
+echo 'updating SQLAlchemy classes...'
+[[ -f $MILDRED_HOME/python/server/db/generated/*.p* ]] && rm $MILDRED_HOME/python/server/db/generated/*.p*
 touch $MILDRED_HOME/python/server/db/__init__.py
 touch $MILDRED_HOME/python/server/db/generated/__init__.py
-sqlacodegen mysql://root:steel@localhost/media > $MILDRED_HOME/python/server/db/generated/sqla_media.py
-sqlacodegen mysql://root:steel@localhost/media > $MILDRED_HOME/python/server/db/generated/sqla_media.py
-# sqlacodegen mysql://root:steel@localhost/analysis --tables dispatch,action,action_param,meta_action,meta_action_param,reason,reason_param,meta_reason,meta_reason_param,action_reason,m_action_m_reason > $MILDRED_HOME/python/server/db/generated/sqla_action.py
-sqlacodegen mysql://root:steel@localhost/analysis > $MILDRED_HOME/python/server/db/generated/sqla_analysis.py
-#sqlacodegen mysql://root:steel@localhost/admin > $MILDRED_HOME/python/server/db/generated/sqla_admin.py
-sqlacodegen mysql://root:steel@localhost/service > $MILDRED_HOME/python/server/db/generated/sqla_service.py
-sqlacodegen mysql://root:steel@localhost/scratch > $MILDRED_HOME/python/server/db/generated/sqla_scratch.py
+
+regenerate_alchemy_classes admin
+regenerate_alchemy_classes analysis
+regenerate_alchemy_classes elastic
+regenerate_alchemy_classes media
+regenerate_alchemy_classes service
+regenerate_alchemy_classes suggestion
+regenerate_alchemy_classes scratch
 
 # cat $MILDRED_HOME/python/server/db/generated/*.py
